+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [Unreleased](https://github.com/ScoopInstaller/Scoop/compare/master...develop)
+### Features
+
+- **scoop-cache:** Add support for CACHEDIR.TAG file ([#6351](https://github.com/ScoopInstaller/Scoop/pull/6351))
+
 ## [v0.5.3](https://github.com/ScoopInstaller/Scoop/compare/v0.5.2...v0.5.3) - 2024-12-31
 
 ### Bug Fixes